requirements.txt
accelerate==1.4.0
addict==2.4.0
av==12.0.0
bitsandbytes==0.45.2
came_pytorch==0.1.3
cuda-python==12.1
decord==0.6.0
deepspeed==0.16.7
diffusers==0.32.1
einops==0.8.1
h5py==3.14.0
huggingface_hub==0.35.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
jsonlines==4.0.0
kornia==0.8.0
matplotlib==3.7.0
moviepy==1.0.3
numpy==1.26.1
omegaconf==2.3.0
opencv-python==4.11.0.86
open3d==0.19.0
plotly==6.2.0
Pillow==10.4.0
prodigyopt==1.1.2
psutil==7.0.0
https://dl.fbaipublicfiles.com/pytorch3d/packaging/wheels/py310_cu121_pyt251/pytorch3d-0.7.8-cp310-cp310-linux_x86_64.whl
pytorch_lightning==2.5.1.post0
PyYAML==6.0.2
requests==2.32.4
safetensors==0.7.0
scikit-learn==1.7.0
scipy==1.15.2
termcolor==3.3.0
tensorboard==2.19.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
https://data.pyg.org/whl/torch-2.5.0%2Bcu121/torch_scatter-2.1.2%2Bpt25cu121-cp310-cp310-linux_x86_64.whl
transformers==4.57.1
trimesh==4.7.0
tqdm==4.67.1
viser==1.0.0
yapf
yq==3.4.3
timm
supervision
pycocotools
volcengine-python-sdk
pydantic
--extra-index-url https://pypi.nvidia.com/
cuml-cu12==23.6